requirements.txt
discord.py
python-dotenv
playwright
python-telegram-bot
nest_asyncio